feature(frontend): Connected the player lobby view to the APIManager - #75
