+++ frontend/src/views/PlayerLobbyView.vue
@@ -2,8 +2,8 @@
     <ScreenBackground blur/>
 
     <CustomButton :action="() => $router.push('/')" style="position: absolute; top: 20px; left: 20px;">Back to Home</CustomButton>
-    <h1>Lobby</h1>
-    <h3>Welcome: ...</h3>
+    <h1>Lobby - {{ lobbyCode }}</h1>
+    <h3>Welcome: {{ playerName }}</h3>
 
     <div class="waiting-message">
         <h2>Waiting for Game </h2>
@@ -18,6 +18,9 @@
 <script lang="ts"  setup>
 import CustomButton from '@/components/CustomButton.vue'
 import ScreenBackground from '@/components/ScreenBackground.vue';
+import router from '@/router';
+import APIManager from '@/types/APIManager';
+import { PlayerSession } from '@/types/PlayerSession';
 import { onMounted, onUnmounted, ref } from 'vue';
 
 const waitingMessages = ref(["to Start", "to Start.", "to Start..", "to Start..."]);
@@ -25,12 +28,30 @@ const currentMessage = ref<string>(waitingMessages.value[0]);
 
 let intervalId: number | undefined;
 
+const lobbyCode = ref('');
+const playerName = ref('');
+
 onMounted(() => {
-  intervalId = setInterval(() => {
+    const session = APIManager.getInstance().getSession();
+    if (!session) {
+        alert('No session found. Please create or join a session first.');
+        router.push('/');
+        return;
+    }
+    if (!(session instanceof PlayerSession)) {
+        alert('This view is only for players. Please switch to the player view.');
+        router.push('/');
+        return;
+    }
+    
+    lobbyCode.value = session.lobbyCode;
+    playerName.value = session.getPlayer().getNickname();
+
+    intervalId = setInterval(() => {
     const currentIndex = waitingMessages.value.indexOf(currentMessage.value);
     const nextIndex = (currentIndex + 1) % waitingMessages.value.length;
     currentMessage.value = waitingMessages.value[nextIndex];
-  }, 500);
+    }, 500);
 });
 
 onUnmounted(() => {
